requirements.txt
aio-pika==9.5.8
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiormq==6.9.2
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncer==0.0.8
asyncpg==0.31.0
attrs==25.4.0
backoff==2.2.1
CacheControl==0.14.4
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorlog==6.10.1
cryptography==46.0.7
diskcache==5.6.3
distro==1.9.0
dspy==3.0.4
Events==0.5
fastapi==0.124.4
fastuuid==0.14.0
filelock==3.20.0
firebase_admin==7.4.0
frozenlist==1.8.0
fsspec==2025.12.0
gepa==0.0.17
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-core==2.5.1
google-cloud-firestore==2.26.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
greenlet==3.3.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
json_repair==0.54.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
litellm==1.80.5
magicattr==0.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.0
numpy==2.3.5
openai==2.12.0
opensearch-protobufs==0.19.0
opensearch-py==3.1.0
optuna==4.6.0
orjson==3.11.5
packaging==25.0
pamqp==3.3.0
pillow==12.0.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.24
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlmodel==0.0.27
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
